reuse cap-std-syncs windows scheduler without copypaste
